backend/pyproject.toml
[tool.poetry]
name = "markdown-manager-api"
version = "0.1.0"
description = "FastAPI backend for Markdown Manager"
authors = ["Dan Little <dan@littledan.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
# Database dependencies
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
aiosqlite = "^0.20.0"
# Authentication
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
# MFA and TOTP
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^8.2"}
# Additional utilities
email-validator = "^2.1.0"
httpx = "^0.28.1"
aiofiles = "^24.1.0"
pygments = "^2.17.0"
premailer = "^3.10.0"
bcrypt = "^4.3.0"
argon2-cffi = "^25.1.0"
pre-commit = "^4.2.0"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
beautifulsoup4 = "^4.13.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
isort = "^5.13.0"
flake8 = "^7.1.0"
mypy = "^1.13.0"
httpx = "^0.28.1"
# Testing utilities
pytest-mock = "^3.12.0"
faker = "^19.6.0"
types-aiofiles = "^24.1.0.20250606"
pytest-json-report = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88


[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = [
    "pydantic.mypy"
]
pretty = true
color_output = true
mypy_path = "typings"
ignore_missing_imports = true
exclude = "tests"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
]

[tool.bandit]
exclude_dirs = []
skips = ["B608"]

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "docker/*",
    "app/main.py",
    "*/migrations/*",
    "*/alembic/*",
    "app/routers/auth_backup.py",
    "app/configs/environment.py",
    "app/services/pdf_processor.py",
    "app/middleware/error_handling.py",
]
branch = true

[tool.coverage.report]
fail_under = 55
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug:",
    "if settings\\.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.xenon]
max_absolute = "B"
max_modules = "B"
max_average = "A"
exclude = ["tests/*", "*/tests/*", "tests/**/*"]
